+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Unreleased
 - lib.platforms.director.ConfigV4: Add struct, checksum generator
 - lib.images.base: Fix keyword usage, naming collision
 - lib.platforms.director.unlock_dir_file: Add helper for unlocking Director files
+- lib.games.lemmings: Replace BitReader with BitStream
 
 0.7.3 - 2019-12-07
 ==================